ai-service/requirements.txt
Django==3.2.19
djangorestframework==3.14.0
psycopg2-binary==2.9.6
requests==2.28.2
gunicorn==21.2.0
python-dotenv==1.0.0
scikit-learn==1.2.2
joblib==1.2.0
numpy==1.23.5  # Changed from 1.24.3 to be compatible with TensorFlow
pandas==2.0.1
tensorflow==2.12.0
transformers==4.28.1
torch==2.0.1
django-cors-headers==4.0.0
dj-database-url==1.2.0